Refactored side cards into one composite with tests and stories
