+++ src/ecrc-frontend/src/components/base/sideCard/SideCard.stories.js
@@ -4,8 +4,9 @@ import SideCard from "./SideCard";
 
 const sideCard = {
   heading: "Get a BC Services Card",
-  content:
-    "B.C. residents who have lived in the province for at least six months must use a BC Services Card to log in to the online qualification tool. Learn how to get a card.",
+  content: [
+    "B.C. residents who have lived in the province for at least six months must use a BC Services Card to log in to the online qualification tool. Learn how to get a card."
+  ],
   type: "blue"
 };
 
@@ -16,8 +17,9 @@ storiesOf("SideCard", module)
       sideCard={{
         ...sideCard,
         heading: "Collection Notice",
-        content:
-          "The information requested on this form is collected by the Liquor and Cannabis Regulation Branch under Sections 26 (a) and (c) of the Freedom of Information and Protection of Privacy Act for the purpose of cannabis licensing, compliance and enforcement matters in accordance with the Cannabis",
+        content: [
+          "The information requested on this form is collected by the Liquor and Cannabis Regulation Branch under Sections 26 (a) and (c) of the Freedom of Information and Protection of Privacy Act for the purpose of cannabis licensing, compliance and enforcement matters in accordance with the Cannabis"
+        ],
         type: "notice"
       }}
     />
@@ -27,7 +29,7 @@ storiesOf("SideCard", module)
       sideCard={{
         ...sideCard,
         heading: "Business Profile",
-        content: "Content for blue grey",
+        content: ["Content for blue grey"],
         type: "bluegrey"
       }}
     />
