+++ src/ecrc-frontend/src/components/composite/SideCards/SideCards.css
@@ -9,4 +9,9 @@ p > button {
 .dark-button {
   background-color: #2d476f;
   border-color: #294266;
+  margin-right: 0px;
+}
+
+.block {
+  display: contents;
 }
