--- src/ecrc-frontend/src/components/composite/accesscodesidecard/AccessCodeSideCard.css
@@ -1,3 +1,0 @@
-.block {
-  display: contents;
-}
