--- src/ecrc-frontend/src/components/composite/accesscodesidecard/AccessCodeSideCard.stories.js
@@ -1,10 +1,0 @@
-import React from "react";
-
-import AccessCardSideCard from "./AccessCodeSideCard";
-
-export default {
-  title: "AccessCardSideCard",
-  component: AccessCardSideCard
-};
-
-export const Default = () => <AccessCardSideCard />;
