--- src/ecrc-frontend/src/components/composite/accesscodesidecard/AccessCodeSideCard.test.js
@@ -1,11 +1,0 @@
-import React from "react";
-import { create } from "react-test-renderer";
-
-import AccessCodeSideCard from "./AccessCodeSideCard";
-
-describe("AccessCodeSideCard Component", () => {
-  test("Matches the snapshot", () => {
-    const accessCode = create(<AccessCodeSideCard />);
-    expect(accessCode.toJSON()).toMatchSnapshot();
-  });
-});
