--- src/ecrc-frontend/src/components/composite/bcservicesidecard/BcServiceSideCard.stories.js
@@ -1,10 +1,0 @@
-import React from "react";
-
-import BcServiceSideCard from "./BcServiceSideCard";
-
-export default {
-  title: "BcServiceSideCard",
-  component: BcServiceSideCard
-};
-
-export const Default = () => <BcServiceSideCard />;
