--- src/ecrc-frontend/src/components/composite/contactinformationsidecard/ContactInformationSideCard.stories.js
@@ -1,10 +1,0 @@
-import React from "react";
-
-import ContactInformationSideCard from "./ContactInformationSideCard";
-
-export default {
-  title: "ContactInformationSideCard",
-  component: ContactInformationSideCard
-};
-
-export const Default = () => <ContactInformationSideCard />;
