--- src/ecrc-frontend/src/components/composite/criminalrecordsidecard/CriminalRecordSIdeCard.test.js
@@ -1,11 +1,0 @@
-import React from "react";
-import { create } from "react-test-renderer";
-
-import CriminalRecordSideCard from "./CriminalRecordSideCard";
-
-describe("CriminalRecordSideCard Component", () => {
-  test("Matches the snapshot", () => {
-    const criminalRecord = create(<CriminalRecordSideCard />);
-    expect(criminalRecord.toJSON()).toMatchSnapshot();
-  });
-});
