--- src/ecrc-frontend/src/components/composite/criminalrecordsidecard/CriminalRecordSideCard.stories.js
@@ -1,10 +1,0 @@
-import React from "react";
-
-import CriminalRecordSideCard from "./CriminalRecordSideCard";
-
-export default {
-  title: "CriminalRecordSideCard",
-  component: CriminalRecordSideCard
-};
-
-export const Default = () => <CriminalRecordSideCard />;
